Move tests to tst. Rename luatest -> tst. Add coverage script.
