--- fmt/fmt_test.go
@@ -1,8 +1,0 @@
-package fmt_test
-
-import (
-	"github.com/Shopify/goluago/luatest"
-	"testing"
-)
-
-func TestLuaFmt(t *testing.T) { luatest.RunLuaTests(t, "fmt_test.lua") }
