+++ net/url/url.go
@@ -5,6 +5,9 @@ import (
 	"net/url"
 )
 
+// Open makes part of the Go net/url package available to Lua code executing
+// in the given lua.State, provided that it requires it:
+//    local url = require("goluago/net/url")
 func Open(l *lua.State) {
 	urlOpen := func(l *lua.State) int {
 		lua.NewLibrary(l, urlLibrary)
