--- net/url/url_test.go
@@ -1,8 +1,0 @@
-package url_test
-
-import (
-	"github.com/Shopify/goluago/luatest"
-	"testing"
-)
-
-func TestLuaURL(t *testing.T) { luatest.RunLuaTests(t, "url_test.lua") }
