--- strings/strings_test.go
@@ -1,8 +1,0 @@
-package strings_test
-
-import (
-	"github.com/Shopify/goluago/luatest"
-	"testing"
-)
-
-func TestLuaStrings(t *testing.T) { luatest.RunLuaTests(t, "strings_test.lua") }
